refactor: moving AppList components into app-list sub-directory.
