+++ src/pages-and-resources/discussions/app-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './AppList';
